cnn/requirements.txt
Flask==2.2.2
keras==2.11.0
numpy==1.20.0
tensorflow==2.11.0
pillow==9.4.0